updated brpel.el for resources and components.
Resources and components can now be viewed and edited.
Currently they still can't be submitted to the BRP server
but this marks the foundation for how navigation works in
the BRP browser.

Components and resources when viewed are now opened in a temporary
json file. This is better than the previous method which enabled
the json major mode manually. Now it takes advantage of the user
configuration for what is to be done when entering a file whose
extension is ".json".